+++ src/async/max.rs
@@ -1,5 +1,5 @@
 use super::*;
-use crate::Result;
+use anyhow::{Context, Result};
 use semver::Version;
 
 /// *__NOTE__ You probably want to use `max_version_async!`*
@@ -35,8 +35,11 @@ pub async fn get_max_version(
     user_agent: &str,
 ) -> Result<Option<Version>> {
     let current_version = Version::parse(current_crate_version)
-        .map_err(|_| "Couldn't parse current version")?;
-    let max_version = get_versions(crate_name, user_agent).await?.max_version;
+        .context("Couldn't parse current version")?;
+    let max_version = get_versions(crate_name, user_agent)
+        .await
+        .context("Couldn't get max version")?
+        .max_version;
     let max_version = if current_version < max_version {
         Some(max_version)
     } else {
@@ -87,8 +90,11 @@ pub async fn get_max_minor_version(
     version: &str,
     user_agent: &str,
 ) -> Result<Option<Version>> {
-    let versions = get_version_list(crate_name, user_agent).await?;
-    let current_version = Version::parse(version).map_err(|_| "Couldn't parse `version`")?;
+    let versions = get_version_list(crate_name, user_agent)
+        .await
+        .context("Couldn't get version list")?;
+    let current_version = Version::parse(version)
+        .context("Couldn't parse `version`")?;
 
     let max_minor_version = versions
         .into_iter()
@@ -140,8 +146,11 @@ pub async fn get_max_patch(
     version: &str,
     user_agent: &str,
 ) -> Result<Option<Version>> {
-    let versions = get_version_list(crate_name, user_agent).await?;
-    let current_version = Version::parse(version).map_err(|_| "Couldn't parse `version`")?;
+    let versions = get_version_list(crate_name, user_agent)
+        .await
+        .context("Couldn't get version list")?;
+    let current_version = Version::parse(version)
+        .context("Couldn't parse `version`")?;
 
     let max_patch = versions
         .into_iter()