+++ src/blocking/mod.rs
@@ -8,7 +8,8 @@
 //! }
 //! ```
 
-use crate::*;
+use anyhow::{Context, Result};
+use crate::{CratesioResponse, Versions};
 pub use max::*;
 pub use newest::*;
 use semver::Version;
@@ -62,12 +63,12 @@ pub fn get_versions(crate_name: &str, user_agent: &str) -> Result<Versions> {
     let response: CratesioResponse = reqwest::blocking::Client::builder()
         .user_agent(format!("{}/{}", crate_name, user_agent))
         .build()
-        .map_err(|_| "Couldn't build client")?
+        .context("Couldn't build client")?
         .get(&url)
         .send()
-        .map_err(|_| "Couldn't request crate info")?
+        .context("Couldn't request crate info")?
         .json()
-        .map_err(|_| "Couldn't parse response to JSON")?;
+        .context("Couldn't read as JSON")?;
 
     Ok(response.versions)
 }
@@ -80,12 +81,12 @@ fn get_version_list(crate_name: &str, user_agent: &str) -> Result<Vec<Version>> 
     let response: CratesioResponse = reqwest::blocking::Client::builder()
         .user_agent(format!("{}/{}", crate_name, user_agent))
         .build()
-        .map_err(|_| "Couldn't build client")?
+        .context("Couldn't build client")?
         .get(&url)
         .send()
-        .map_err(|_| "Couldn't request crate info")?
+        .context("Couldn't request crate info")?
         .json()
-        .map_err(|_| "Couldn't parse response to JSON")?;
+        .context("Couldn't read as JSON")?;
     let versions = response.all_versions;
     let versions = versions
         .into_iter()